Update engage_parser.py
Fixed problems with the parser checking tokens and marking them incorrectly. Confirmed as a problem with "is" being marked as a keyword instead of an operator when using code like "is greater than", thus failing to run and causing the parser to expect a "then" in the wrong context point of the token list.